+++ example/src/Screens/LinkComponent.tsx
@@ -2,7 +2,6 @@ import { Button } from '@react-navigation/elements';
 import {
   CommonActions,
   Link,
-  type ParamListBase,
   type PathConfigMap,
   StackActions,
 } from '@react-navigation/native';
@@ -22,15 +21,15 @@ export type LinkComponentDemoParamList = {
   Albums: undefined;
 };
 
-export const linkComponentDemoLinking: PathConfigMap<LinkComponentDemoParamList> =
-  {
-    Article: COMMON_LINKING_CONFIG.Article,
-    Albums: 'albums',
-  };
+const linking: PathConfigMap<LinkComponentDemoParamList> = {
+  Article: COMMON_LINKING_CONFIG.Article,
+  Albums: 'albums',
+};
 
 const scrollEnabled = Platform.select({ web: true, default: false });
 
 const ArticleScreen = ({
+  navigation,
   route,
 }: StackScreenProps<LinkComponentDemoParamList, 'Article'>) => {
   return (
@@ -46,12 +45,35 @@ const ArticleScreen = ({
         >
           Replace with Albums
         </Link>
+
         <Button screen="Home" variant="filled">
           Go to Home
         </Button>
         <Button variant="tinted" action={CommonActions.goBack()}>
           Go back
         </Button>
+
+        <Button
+          variant="tinted"
+          onPress={() =>
+            navigation.setParams({
+              author:
+                route.params?.author === 'Gandalf' ? 'Babel fish' : 'Gandalf',
+            })
+          }
+        >
+          Update params
+        </Button>
+
+        {Platform.OS === 'web' && (
+          <Button
+            onPress={() => {
+              location.hash = 'frodo';
+            }}
+          >
+            Add hash to URL
+          </Button>
+        )}
       </View>
       <Article
         author={{ name: route.params.author }}
@@ -91,18 +113,9 @@ const AlbumsScreen = ({
 
 const SimpleStack = createStackNavigator<LinkComponentDemoParamList>();
 
-type Props = Partial<React.ComponentProps<typeof SimpleStack.Navigator>> &
-  StackScreenProps<ParamListBase>;
-
-export function LinkComponent({ navigation, ...rest }: Props) {
-  React.useLayoutEffect(() => {
-    navigation.setOptions({
-      headerShown: false,
-    });
-  }, [navigation]);
-
+export function LinkComponent() {
   return (
-    <SimpleStack.Navigator {...rest}>
+    <SimpleStack.Navigator>
       <SimpleStack.Screen
         name="Article"
         component={ArticleScreen}
@@ -119,6 +132,9 @@ export function LinkComponent({ navigation, ...rest }: Props) {
     </SimpleStack.Navigator>
   );
 }
+
+LinkComponent.title = '<Link /> ';
+LinkComponent.linking = linking;
 
 const styles = StyleSheet.create({
   buttons: {