+++ packages/bottom-tabs/src/TransitionConfigs/TransitionPresets.tsx
@@ -1,13 +1,13 @@
 import type { BottomTabTransitionPreset } from '../types';
-import { forCrossFade, forShifting } from './SceneStyleInterpolators';
-import { CrossFadeSpec, ShiftingSpec } from './TransitionSpecs';
+import { forFade, forShift } from './SceneStyleInterpolators';
+import { FadeSpec, ShiftSpec } from './TransitionSpecs';
 
 export const FadeTransition: BottomTabTransitionPreset = {
-  transitionSpec: CrossFadeSpec,
-  sceneStyleInterpolator: forCrossFade,
+  transitionSpec: FadeSpec,
+  sceneStyleInterpolator: forFade,
 };
 
-export const ShiftingTransition: BottomTabTransitionPreset = {
-  transitionSpec: ShiftingSpec,
-  sceneStyleInterpolator: forShifting,
+export const ShiftTransition: BottomTabTransitionPreset = {
+  transitionSpec: ShiftSpec,
+  sceneStyleInterpolator: forShift,
 };