+++ packages/core/src/__tests__/StaticNavigation.test.tsx
@@ -7,7 +7,11 @@ import * as React from 'react';
 
 import { BaseNavigationContainer } from '../BaseNavigationContainer';
 import { createNavigatorFactory } from '../createNavigatorFactory';
-import { createComponentForStaticNavigation } from '../StaticNavigation';
+import { getStateFromPath } from '../getStateFromPath';
+import {
+  createComponentForStaticNavigation,
+  createPathConfigForStaticNavigation,
+} from '../StaticNavigation';
 import type { EventMapBase } from '../types';
 import { useIsFocused } from '../useIsFocused';
 import { useNavigationBuilder } from '../useNavigationBuilder';
@@ -232,11 +236,182 @@ it('handles non-function screens', () => {
   }).not.toThrow();
 });
 
-it("throws if screens property isn't specified", () => {
+it("throws if screens or groups property isn't specified", () => {
   expect(() => {
     // @ts-expect-error: we're testing invalid input here
     const Root = createTestNavigator({});
 
     createComponentForStaticNavigation(Root, 'RootNavigator');
-  }).toThrow("Couldn't find a 'screens' property");
+  }).toThrow("Couldn't find a 'screens' or 'groups' property");
+});
+
+it("doesn't throw if either screens or groups property is specified", () => {
+  expect(() => {
+    const Root = createTestNavigator({
+      screens: {},
+    });
+
+    createComponentForStaticNavigation(Root, 'RootNavigator');
+  }).not.toThrow();
+
+  expect(() => {
+    const Root = createTestNavigator({
+      groups: {},
+    });
+
+    createComponentForStaticNavigation(Root, 'RootNavigator');
+  }).not.toThrow();
+});
+
+it('creates linking configuration for static config', () => {
+  const Nested = createTestNavigator({
+    screens: {
+      Profile: {
+        screen: TestScreen,
+        linking: {
+          path: 'profile/:id',
+          parse: {
+            id: Number,
+          },
+        },
+      },
+      Settings: {
+        screen: TestScreen,
+        options: {
+          testId: 'settings',
+        },
+        linking: {
+          path: 'settings',
+          exact: true,
+        },
+      },
+    },
+    groups: {
+      Auth: {
+        screens: {
+          Login: {
+            screen: TestScreen,
+            linking: 'login',
+          },
+          Register: {
+            screen: TestScreen,
+            linking: 'register',
+          },
+          Forgot: {
+            screen: TestScreen,
+            linking: {
+              path: 'forgot-password',
+              exact: true,
+            },
+          },
+        },
+      },
+    },
+  });
+
+  const Root = createTestNavigator({
+    screens: {
+      Home: TestScreen,
+      Feed: {
+        screen: TestScreen,
+        linking: 'feed',
+      },
+      Nested: {
+        screen: Nested,
+        linking: 'nested',
+      },
+    },
+    groups: {
+      Support: {
+        screens: {
+          Contact: {
+            screen: TestScreen,
+            linking: 'contact',
+          },
+          FAQ: {
+            screen: TestScreen,
+            linking: 'faq',
+          },
+        },
+      },
+    },
+  });
+
+  const screens = createPathConfigForStaticNavigation(Root);
+
+  expect(screens).toMatchInlineSnapshot(`
+{
+  "Contact": {
+    "path": "contact",
+  },
+  "FAQ": {
+    "path": "faq",
+  },
+  "Feed": {
+    "path": "feed",
+  },
+  "Nested": {
+    "path": "nested",
+    "screens": {
+      "Forgot": {
+        "exact": true,
+        "path": "forgot-password",
+      },
+      "Login": {
+        "path": "login",
+      },
+      "Profile": {
+        "parse": {
+          "id": [Function],
+        },
+        "path": "profile/:id",
+      },
+      "Register": {
+        "path": "register",
+      },
+      "Settings": {
+        "exact": true,
+        "path": "settings",
+      },
+    },
+  },
+}
+`);
+
+  expect(getStateFromPath('contact', { screens })).toEqual({
+    routes: [
+      {
+        name: 'Contact',
+        path: 'contact',
+      },
+    ],
+  });
+
+  expect(getStateFromPath('settings', { screens })).toEqual({
+    routes: [
+      {
+        name: 'Nested',
+        state: {
+          routes: [{ name: 'Settings', path: 'settings' }],
+        },
+      },
+    ],
+  });
+
+  expect(getStateFromPath('nested/profile/123', { screens })).toEqual({
+    routes: [
+      {
+        name: 'Nested',
+        state: {
+          routes: [
+            {
+              name: 'Profile',
+              path: 'nested/profile/123',
+              params: { id: 123 },
+            },
+          ],
+        },
+      },
+    ],
+  });
 });