+++ packages/devtools/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-alpha.13](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.12...@react-navigation/devtools@7.0.0-alpha.13) (2024-03-09)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+# [7.0.0-alpha.12](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.11...@react-navigation/devtools@7.0.0-alpha.12) (2024-03-08)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+# [7.0.0-alpha.11](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.10...@react-navigation/devtools@7.0.0-alpha.11) (2024-03-04)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
 # [7.0.0-alpha.10](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.9...@react-navigation/devtools@7.0.0-alpha.10) (2024-02-24)
 
 ### Bug Fixes