+++ packages/react-native-tab-view/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-alpha.7](https://github.com/react-navigation/react-navigation/compare/react-native-tab-view@4.0.0-alpha.6...react-native-tab-view@4.0.0-alpha.7) (2024-03-08)
+
+### Bug Fixes
+
+* add href to tabs in material top tabs ([#11873](https://github.com/react-navigation/react-navigation/issues/11873)) ([c3459e1](https://github.com/react-navigation/react-navigation/commit/c3459e10e56446554d45130b28b232f08f491c4b)) - by @groot007
+
+### Features
+
+* implement tab-view new api ([#11548](https://github.com/react-navigation/react-navigation/issues/11548)) ([dca15c9](https://github.com/react-navigation/react-navigation/commit/dca15c9126f8751cfea43edc80c51d28de8f6fa6)) - by @okwasniewski
+
+### BREAKING CHANGES
+
+* react-native-tab-view now has a new API to address performance issues with current
+implementation.
+
+Co-authored-by: Michał Osadnik <micosa97@gmail.com>
+Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>
+
+# [4.0.0-alpha.6](https://github.com/react-navigation/react-navigation/compare/react-native-tab-view@4.0.0-alpha.5...react-native-tab-view@4.0.0-alpha.6) (2024-03-04)
+
+### Bug Fixes
+
+* update drawer and material tab bar to match latest md guidelines ([#11864](https://github.com/react-navigation/react-navigation/issues/11864)) ([8726597](https://github.com/react-navigation/react-navigation/commit/872659710dec1b097ec7c7b1dd59a6174e021b30)) - by @groot007
+
 # [4.0.0-alpha.5](https://github.com/react-navigation/react-navigation/compare/react-native-tab-view@4.0.0-alpha.4...react-native-tab-view@4.0.0-alpha.5) (2024-02-24)
 
 ### Bug Fixes